Separate safe math functions from linearAlgebra.h
