+++ src/architecture/utilities/safeMath.c
@@ -1,0 +1,43 @@
+/*
+ ISC License
+
+ Copyright (c) 2025, Laboratory for Atmospheric and Space Physics, University of Colorado at Boulder
+
+ Permission to use, copy, modify, and/or distribute this software for any
+ purpose with or without fee is hereby granted, provided that the above
+ copyright notice and this permission notice appear in all copies.
+
+ THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+ WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+ MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+
+ */
+
+#include "safeMath.h"
+
+#include <math.h>
+
+double safeAcos(double x) {
+    if (x < -1.0)
+        return acos(-1);
+    else if (x > 1.0)
+        return acos(1);
+    return acos(x);
+}
+
+double safeAsin(double x) {
+    if (x < -1.0)
+        return asin(-1);
+    else if (x > 1.0)
+        return asin(1);
+    return asin(x);
+}
+
+double safeSqrt(double x) {
+    if (x < 0.0) return 0.0;
+    return sqrt(x);
+}
