+++ src/fswAlgorithms/attDetermination/CSSEst/cssWlsEst.cpp
@@ -20,6 +20,7 @@
 #include "fswAlgorithms/attDetermination/CSSEst/cssWlsEst.h"
 #include "architecture/utilities/linearAlgebra.h"
 #include "architecture/utilities/macroDefinitions.h"
+#include "architecture/utilities/safeMath.h"
 #include <string.h>
 
 int computeWlsmn(int numActiveCss, double *H, double *W, double *y, double x[3]);
