+++ src/fswAlgorithms/attGuidance/locationPointing/locationPointing.cpp
@@ -20,8 +20,10 @@
 
 #include "fswAlgorithms/attGuidance/locationPointing/locationPointing.h"
 #include "architecture/utilities/linearAlgebra.h"
-#include "architecture/utilities/rigidBodyKinematics.h"
 #include "architecture/utilities/macroDefinitions.h"
+#include "architecture/utilities/rigidBodyKinematics.h"
+#include "architecture/utilities/safeMath.h"
+
 #include <math.h>
 
 /*! This method performs a complete reset of the module.  Local module variables that retain
