+++ src/fswAlgorithms/formationFlying/etSphericalControl/etSphericalControl.cpp
@@ -16,24 +16,16 @@
  OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
  */
-/*
-    FSW Electrostatic Tractor Control
 
- */
-
-/* modify the path to reflect the new module names */
 #include "etSphericalControl.h"
 
-#include <math.h>
-#include <string.h>
-
-/*
- Pull in support files from other modules.  Be sure to use the absolute path relative to Basilisk directory.
- */
-#include "architecture/utilities/macroDefinitions.h"
 #include "architecture/utilities/linearAlgebra.h"
+#include "architecture/utilities/macroDefinitions.h"
 #include "architecture/utilities/orbitalMotion.h"
 #include "architecture/utilities/rigidBodyKinematics.h"
+#include "architecture/utilities/safeMath.h"
+
+#include <math.h>
 
 
 /*! This method performs a complete reset of the module.  Local module variables that retain
