+++ src/simulation/environment/eclipse/eclipse.cpp
@@ -18,9 +18,11 @@
  */
 
 #include "eclipse.h"
-#include <iostream>
+
 #include "architecture/utilities/astroConstants.h"
 #include "architecture/utilities/avsEigenSupport.h"
+#include "architecture/utilities/linearAlgebra.h"
+#include "architecture/utilities/safeMath.h"
 
 
 Eclipse::Eclipse()
