+++ src/simulation/environment/groundMapping/groundMapping.cpp
@@ -22,6 +22,7 @@
 #include "architecture/utilities/avsEigenSupport.h"
 #include "architecture/utilities/linearAlgebra.h"
 #include "architecture/utilities/rigidBodyKinematics.h"
+#include "architecture/utilities/safeMath.h"
 #include <iostream>
 #include <math.h>
