+++ src/simulation/environment/magneticFieldWMM/magneticFieldWMM.cpp
@@ -20,6 +20,7 @@
 #include "magneticFieldWMM.h"
 #include "architecture/utilities/linearAlgebra.h"
 #include "architecture/utilities/rigidBodyKinematics.h"
+#include "architecture/utilities/safeMath.h"
 #include "EGM9615.h"
 
 /*! The constructor method initializes the dipole parameters to zero, resuling in a zero magnetic field result by default.
