+++ src/simulation/navigation/pinholeCamera/pinholeCamera.cpp
@@ -18,9 +18,11 @@
  */
 
 #include "pinholeCamera.h"
+#include "architecture/utilities/astroConstants.h"
 #include "architecture/utilities/avsEigenSupport.h"
 #include "architecture/utilities/linearAlgebra.h"
 #include "architecture/utilities/rigidBodyKinematics.h"
+#include "architecture/utilities/safeMath.h"
 #include <iostream>
 #include <math.h>
